requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Flask-Login==0.6.2
gunicorn==20.1.0
pandas==2.0.3
python-dotenv==1.0.0